+++ R/class-worksheet.R
@@ -137,7 +137,7 @@ wbWorksheet <- R6::R6Class(
 
     #' @description
     #' Creates a new `wbWorksheet` object
-    #' @param tabColour tabColour
+    #' @param tabColor tabColor
     #' @param oddHeader oddHeader
     #' @param oddFooter oddFooter
     #' @param evenHeader evenHeader
@@ -151,7 +151,7 @@ wbWorksheet <- R6::R6Class(
     #' @param printGridLines printGridLines
     #' @return a `wbWorksheet` object
     initialize = function(
-      tabColour   = NULL,
+      tabColor   = NULL,
       oddHeader   = NULL,
       oddFooter   = NULL,
       evenHeader  = NULL,
@@ -164,10 +164,10 @@ wbWorksheet <- R6::R6Class(
       vdpi        = 300,
       printGridLines = FALSE
     ) {
-      if (!is.null(tabColour)) {
-        tabColour <- sprintf('<sheetPr><tabColor rgb="%s"/></sheetPr>', tabColour)
+      if (!is.null(tabColor)) {
+        tabColor <- sprintf('<sheetPr><tabColor rgb="%s"/></sheetPr>', tabColor)
       } else {
-        tabColour <- character()
+        tabColor <- character()
       }
 
       hf <- list(
@@ -189,7 +189,7 @@ wbWorksheet <- R6::R6Class(
       }
 
       ## list of all possible children
-      self$sheetPr               <- tabColour
+      self$sheetPr               <- tabColor
       self$dimension             <- '<dimension ref="A1"/>'
       self$sheetViews            <- character()
       self$sheetFormatPr         <- '<sheetFormatPr baseColWidth="8.43" defaultRowHeight="16" x14ac:dyDescent="0.2"/>'