+++ tests/testthat/test-base_font.R
@@ -5,7 +5,7 @@ test_that("get_base_font works", {
     list(
       size = list(val = "11"),
       # should this be "#000000"?
-      colour = list(theme = "1"),
+      color = list(theme = "1"),
       name = list(val = "Calibri")
     )
   )
@@ -15,7 +15,7 @@ test_that("get_base_font works", {
     wb$get_base_font(),
     list(
       size = list(val = "9"),
-      colour = list(rgb = "FFFF0000"),
+      color = list(rgb = "FFFF0000"),
       name = list(val = "Arial")
     )
   )